+++ computer_vision/object_detection/yolo/main.py
@@ -39,7 +39,7 @@ def detect_chairs(filepath="data/chair.jpg"):
     boxes = []
     for category, score, bounds in results:
         if category == "chair":
-            x, y, w, h = bounds
+            x, y, w, h = bounds  # Midpoint x and y, width and height
             boxes.append([x, y, w, h])
     return boxes
 
@@ -52,9 +52,23 @@ def visualisation(filepath="data/chair.jpg", writeto="detected.jpg"):
         thickness = 5 if RESIZE else 10
         font_scale = w / 100
         font_thickness = 1 if RESIZE else 5
-        cv2.rectangle(img, (int(x - w / 2), int(y - h / 2)), (int(x + w / 2), int(y + h / 2)), BLUE, thickness=thickness)
+        cv2.rectangle(img, (int(x - (w / 2)), int(y - (h / 2))), (int(x + (w / 2)), int(y + (h / 2))), BLUE, thickness=thickness)
         cv2.putText(img, "chair", (int(x),int(y)), cv2.FONT_HERSHEY_PLAIN, font_scale, BLUE, font_thickness)
     cv2.imwrite(writeto, img)
+
+def get_box_corners(filepath="data/chair.jpg"):
+    """Returns a list of coordinates of the top left and bottom right corners of bounding boxes for an image."""
+    boxes = detect_chairs(filepath)
+    corners = []
+    for box in boxes:
+        x, y, w, h = box
+        upper_left_x = x - (w / 2)
+        upper_left_y = y - (h / 2)
+        lower_right_x = x + (w / 2)
+        lower_right_y = y + (h / 2)
+        corners.append([(upper_left_x, upper_left_y), (lower_right_x, lower_right_y)])
+    return corners
+
 
 # Test Outputs
 visualisation()